requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
fastapi-sessions==0.3.2
fasttext==0.9.2
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.39.4   
frozenlist==1.3.3
fsspec==2023.5.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.89.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.26.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-documentai==2.15.0
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
orjson==3.9.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
preshed==3.0.8
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Shapely==1.8.5.post1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
thinc==8.1.10
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wasabi==1.1.1
watermark==2.4.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
